+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.1.0
+
+* updateItemAt method was implemented.
+* removeItemAt method was implemented.
+* another improvements.
+
 ## 0.0.1
 
 * First release